refactor: rename directory back to fix git issue
