+++ extensions/bitwarden/src/components/generatePassword/ActionPanel.tsx
