+++ extensions/bitwarden/src/components/generatePassword/OptionField.tsx
